# product-service/swagger.yml
swagger: '2.0'
info:
  description: |
    This is Swagger schema for RS Node AWS
  version: 1.0.0
  title: Swagger PS4 Games Store
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# basePath: /v2
tags:
  - name: products
    description: PS4 Games
# schemes:
# - http
paths:
  /products:
    get:
      tags:
        - products
      summary: Get List of all existing PS4 games in store
      operationId: getProductsList
      produces:
        - application/json
        - application/xml
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        500:
          description: Internal Server Error
    post:
      tags:
        - products
      summary: Create new product
      operationId: createProduct
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Product object that needs to be added to the store
          required: true
          schema:
            $ref: '#/definitions/NewProduct'
      responses:
        400:
          description: 'Error: Bad request'
  /products/{productId}:
    get:
      tags:
        - products
      summary: Find PS4 game by ID
      description: Returns a single product
      operationId: getProductsById
      produces:
        - application/json
        - application/xml
      parameters:
        - name: productId
          in: path
          description: ID of game to return
          required: true
          type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Product'
        404:
          description: Game not found
        500:
          description: Internal Server Error
definitions:
  Product:
    type: object
    required:
      - id
      - count
      - description
      - price
      - title
    properties:
      id:
        type: string
      count:
        type: integer
        format: int64
      description:
        type: string
      price:
        type: number
      title:
        type: string
  NewProduct:
    type: object
    required:
      - count
      - description
      - price
      - title
    properties:
      count:
        type: integer
        format: int64
      description:
        type: string
      price:
        type: number
      title:
        type: string
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
# Added by API Auto Mocking Plugin
# host: 64rj2xj4v3.execute-api.eu-west-1.amazonaws.com
# basePath: /dev
schemes:
  - https
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /DzmitryULearn/rs-aws-be/1.0.0